requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
tensorflow==2.15.0
joblib==1.4.0
Pillow==10.3.0
matplotlib==3.8.4